# openapi.yaml
openapi: 3.0.1
info:
  title: My API
  description: Документация к API проекта CompClub
  version: v1
paths:
...
  /api/Account/get_info/{id}:
    get:
      tags:
        - Account
      summary: Получение информации об аккаунте по его ID (только для администратора).
      description: Возвращает подробную информацию об аккаунте.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/AccountGetInfoResponseSchema'
            application/json:
              schema:
                $ref: '#/components/schemas/AccountGetInfoResponseSchema'
            text/json:
              schema:
                $ref: '#/components/schemas/AccountGetInfoResponseSchema'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/Account/get_info:
    get:
      tags:
        - Account
      summary: Получение информации о своем аккаунте (для клиента).
      description: Авторизованный клиент может получить свои данные.
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Account'
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
            text/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/Account/update/{id}:
    put:
      tags:
        - Account
      summary: Обновление информации об аккаунте по ID (только для администратора или владельца).
      description: Можно частично обновить данные (логин, email, пароль).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountByIdModel'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountByIdModel'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdateAccountByIdModel'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MessageResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/Account/update:
    put:
      tags:
        - Account
      summary: Обновление своей информации (для клиента).
      description: Клиент может изменить свой логин и пароль.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountModel'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountModel'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdateAccountModel'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MessageResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/Account/add_balance:
    post:
      tags:
        - Account
      summary: Пополнение баланса текущего аккаунта (только для администратора или владельца).
      description: Добавляет указанную сумму к текущему балансу и сохраняет историю.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddBalanceModel'
          text/json:
            schema:
              $ref: '#/components/schemas/AddBalanceModel'
          application/*+json:
            schema:
              $ref: '#/components/schemas/AddBalanceModel'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MessageResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/Account/add_balance_by_id/{id}:
    post:
      tags:
        - Account
      summary: Пополнение баланса аккаунта по ID (только для администратора или владельца).
      description: Также добавляется запись о доходе для клуба.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: idClub
          in: query
          schema:
            type: integer
            format: int32
        - name: money
          in: query
          schema:
            type: number
            format: double
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MessageResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/Account/change_password:
    post:
      tags:
        - Account
      summary: Смена пароля для авторизованного клиента.
      description: Хэширует новый пароль и сохраняет его.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordModel'
          text/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordModel'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ChangePasswordModel'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MessageResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/Account/change_password_by_email/{email}:
    post:
      tags:
        - Account
      summary: Сброс пароля по email.
      description: Генерирует временный пароль и отправляет его на указанный email.
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MessageResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/Account/deactivate_account/{id}:
    put:
      tags:
        - Account
      summary: Деактивация аккаунта по ID (только для администратора или владельца).
      description: Делает аккаунт неактивным без удаления.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MessageResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/Account/authentication:
    post:
      tags:
        - Account
      summary: Аутентификация пользователя и выдача JWT-токена.
      description: Проверяет логин и пароль, возвращает JWT при успешной аутентификации.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthModel'
          text/json:
            schema:
              $ref: '#/components/schemas/AuthModel'
          application/*+json:
            schema:
              $ref: '#/components/schemas/AuthModel'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/TokenResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    Account:
      type: object
      properties:
        id:
          type: integer
          format: int32
        idClient:
          type: integer
          format: int32
        balance:
          type: number
          format: double
          nullable: true
        login:
          type: string
          nullable: true
        lastLogin:
          type: string
          format: date-time
          nullable: true
        passwordChangedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
        email:
          type: string
          nullable: true
        isAlive:
          type: boolean
          nullable: true
        idClientNavigation:
          $ref: '#/components/schemas/Client'
      additionalProperties: false
    AccountGetInfoResponseSchema:
      type: object
      properties:
        account:
          $ref: '#/components/schemas/Account'
      additionalProperties: false
    AddBalanceModel:
      type: object
      properties:
        money:
          type: number
          format: double
      additionalProperties: false
    AuthModel:
      type: object
      properties:
        login:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
      additionalProperties: false
...
    ChangePasswordModel:
      type: object
      properties:
        password:
          type: string
          nullable: true
      additionalProperties: false
    Client:
      type: object
      properties:
        id:
          type: integer
          format: int32
        firstName:
          type: string
          nullable: true
        middleName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          nullable: true
      additionalProperties: false
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          nullable: true
      additionalProperties: false
...
    TokenResponse:
      type: object
      properties:
        token:
          type: string
          nullable: true
      additionalProperties: false
    UpdateAccountByIdModel:
      type: object
      properties:
        login:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
      additionalProperties: false
    UpdateAccountModel:
      type: object
      properties:
        login:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: apiKey
      description: Введите 'Bearer' [пробел] и ваш токен в поле ниже для доступа к защищенным ресурсам
      name: Authorization
      in: header
security:
  - Bearer: []